WIP(AnalyticalTable): test updates
